feat: 合并新建/编辑TODO为统一抽屉组件 (#302)
将任务新建(CreateTodoModal)、编辑(TodoEditDrawer)和设置(TodoSettingsDrawer)
三个分散的界面合并为统一的 TodoDrawer 组件，实现:
- 统一的布局和交互体验
- 一步完成创建/编辑操作
- 合并了标题、Prompt、执行器、调度、工作目录等所有设置

BREAKING CHANGE: 移除 CreateTodoModal、TodoEditDrawer、TodoSettingsDrawer 组件
